+++ lib/Settings/Admin.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * SPDX-FileCopyrightText: 2020 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
@@ -6,6 +7,7 @@
 namespace OCA\Github\Settings;
 
 use OCA\Github\AppInfo\Application;
+use OCA\Github\Service\GithubAPIService;
 use OCA\Github\Service\SecretService;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Services\IInitialState;
@@ -18,6 +20,7 @@ class Admin implements ISettings {
 		private SecretService $secretService,
 		private IConfig $config,
 		private IInitialState $initialStateService,
+		private GitHubAPIService $githubAPIService,
 	) {
 	}
 
@@ -31,8 +34,19 @@ class Admin implements ISettings {
 		$adminDashboardEnabled = $this->config->getAppValue(Application::APP_ID, 'dashboard_enabled', '1') === '1';
 		$adminLinkPreviewEnabled = $this->config->getAppValue(Application::APP_ID, 'link_preview_enabled', '1') === '1';
 		$defaultLinkToken = $this->secretService->getEncryptedAppValue('default_link_token');
+		$userName = $this->config->getAppValue(Application::APP_ID, 'user_name');
+		$userDisplayName = $this->config->getAppValue(Application::APP_ID, 'user_displayname');
 		$allowDefaultTokenToAnonymous = $this->config->getAppValue(Application::APP_ID, 'allow_default_link_token_to_anonymous', '0') === '1';
 		$allowDefaultTokenToGuests = $this->config->getAppValue(Application::APP_ID, 'allow_default_link_token_to_guests', '0') === '1';
+		$issueNotificationsEnabled = $this->config->getAppValue(Application::APP_ID, 'issue_notifications_enabled', '1') === '1';
+
+		if ($defaultLinkToken !== '' && ($userName === '' || $userDisplayName === '')) {
+			$userInfo = $this->githubAPIService->updateLinkTokenUserInfo($defaultLinkToken);
+			if (isset($userInfo['user_name'], $userInfo['user_displayname'])) {
+				$userName = $userInfo['user_name'];
+				$userDisplayName = $userInfo['user_displayname'];
+			}
+		}
 
 		$adminConfig = [
 			'client_id' => $clientID,
@@ -43,6 +57,9 @@ class Admin implements ISettings {
 			'default_link_token' => $defaultLinkToken === '' ? '' : 'dummyToken',
 			'allow_default_link_token_to_anonymous' => $allowDefaultTokenToAnonymous,
 			'allow_default_link_token_to_guests' => $allowDefaultTokenToGuests,
+			'user_name' => $userName,
+			'user_displayname' => $userDisplayName,
+			'issue_notifications_enabled' => $issueNotificationsEnabled,
 		];
 		$this->initialStateService->provideInitialState('admin-config', $adminConfig);
 		return new TemplateResponse(Application::APP_ID, 'adminSettings');