+++ CHANGELOG.md
@@ -8,6 +8,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.2.0 - 2025-05-19
+
+Maintenance update
+
+### Added
+
+- Added default access token user info in Admin settings
+- Added notifications for new unread GitHub notifications (admin/user option to enable/disable)
+
+
 ## 3.1.1 – 2024-11-06
 
 ### Changed